Fix confusingly named command line argument
--allowed-kimera is confusing; does kimera refer to a chimeric read or does it
refer to a kmer? Clearly, when following the code path in core/pipeline.py, one
sees that it is used as the argument to the --kmer switch for TagGD, so I guess
we should reflect this also in the pipeline's CLI.
Note that for subsequent releases the switches in sbatch scripts will have to
be fixed with respect to this!

Change-Id: I77b015d9bda8192539ee3c0123afa479a9e9975d