Base versioning on git tags
The function get_git_version() reads the version information from the
file st_pipeline/pipeline/RELEASE-VERSION. It also tries to get it via
'git describe'. If the latter succeeds and two differ it will update the
file. What we will need in addition to this is a simple script to make a
release (which Jose informs me is just a matter of zip'ing the
directory) to ensure that this function is runs before any release is
made, and thus to guarantee that the RELEASE-VERSION is correct in any
released package. Right now, the function get_git_version() is run in
the initialization code for the stpipeline package, i.e. in the
__init__.py file.

Change-Id: I47f4971afc0298de9b5d2d95d7d77cd7e0b21939